requirements.txt
langchain_openai
langchain
langchain_core
langchain_community
langchain_ollama
python-dotenv
streamlit
openai

